test: move safetarfile to another test module
